rename composer to html-composer
