+++ lib/html-composer.js
