+++ lib/html-composer.test.js
