+++ lib/routes.js
@@ -10,7 +10,7 @@
 const _ = require('lodash'),
   express = require('express'),
   siteService = require('./services/sites'),
-  composer = require('./composer'),
+  htmlComposer = require('./html-composer'),
   responses = require('./responses'),
   files = require('./files'),
   sitemaps = require('./sitemaps'),
@@ -129,7 +129,7 @@ function addSiteController(router, site) {
 
     if (controller) {
       if (_.isFunction(controller)) {
-        controller(router, composer, site);
+        controller(router, htmlComposer, site);
       }
 
       // things to remember from controller
