+++ lib/routes/components.js
@@ -8,7 +8,7 @@
 
 var _ = require('lodash'),
   responses = require('../responses'),
-  composer = require('../composer'),
+  htmlComposer = require('../html-composer'),
   files = require('../files'),
   queryStringOptions = ['ignore-data'],
   controller = require('../services/components'),
@@ -132,7 +132,7 @@ let route = _.bindAll({
 
   render: function (req, res) {
     responses.expectHTML(function () {
-      return composer.renderComponent(req.uri, res, _.pick(req.query, queryStringOptions));
+      return htmlComposer.renderComponent(req.uri, res, _.pick(req.query, queryStringOptions));
     }, res);
   }
 });
