+++ lib/routes/pages.js
@@ -8,7 +8,7 @@
 
 var _ = require('lodash'),
   responses = require('../responses'),
-  composer = require('../composer'),
+  htmlComposer = require('../html-composer'),
   controller = require('../services/pages'),
   acceptedExtensions = {
     html: 'text/html',
@@ -73,7 +73,7 @@ let route = _.bindAll({
    */
   render: function (req, res) {
     responses.expectHTML(function () {
-      return composer.renderPage(req.uri, res);
+      return htmlComposer.renderPage(req.uri, res);
     }, res);
   }
 });
