+++ lib/setup.js
@@ -3,7 +3,7 @@
 var express = require('express'),
   routes = require('./routes'),
   bootstrap = require('./bootstrap'),
-  composer = require('./composer');
+  htmlComposer = require('./html-composer');
 
 /**
  * optionally pass in an express app and/or templating engines
@@ -24,7 +24,7 @@ module.exports = function (options) {
 
   // if engines were passed in, send them to the composer
   if (engines) {
-    composer.addEngines(engines);
+    htmlComposer.addEngines(engines);
   }
 
   //look for bootstraps in components
