+++ lib/setup.test.js
@@ -4,14 +4,14 @@ var _ = require('lodash'),
   filename = __filename.split('/').pop().split('.').shift(),
   lib = require('./' + filename),
   sinon = require('sinon'),
-  composer = require('./composer');
+  htmlComposer = require('./html-composer');
 
 describe(_.startCase(filename), function () {
   var sandbox;
 
   beforeEach(function () {
     sandbox = sinon.sandbox.create();
-    sandbox.stub(composer, 'addEngines');
+    sandbox.stub(htmlComposer, 'addEngines');
   });
 
   afterEach(function () {
@@ -24,7 +24,7 @@ describe(_.startCase(filename), function () {
 
   it('adds engines', function () {
     return lib({engines: 'whatever'}).then(function () {
-      sinon.assert.calledOnce(composer.addEngines);
+      sinon.assert.calledOnce(htmlComposer.addEngines);
     });
   });
 });
